before adding simple BPSK modulation

+++ SonicPACT-app/src/main/cpp/AudoGenerator.h
@@ -54,7 +54,7 @@ private:
     // Wave params, these could be instance variables in order to modify at runtime
     static float constexpr kAmplitude = 0.5f;
 //    static float constexpr mListenFrequency = 21000;
-    static float constexpr kPI = M_PI;
+    static float constexpr kPI = static_cast<const float>(M_PI);
     static float constexpr kTwoPi = kPI * 2;
     bool shouldBroadcast = false;
     bool isBroadcasting = false;
